fix: consolidate HTTP transport package files
